refactor: move test helper funcs
